requirements.txt
Flask==3.1.2
spotipy==2.25.1
requests==2.32.5
python-dotenv==1.0.0
